@@ -38,10 +38,11 @@ namespace TwitchAchievementTrackerBackend.Controllers
             {
                 var xConfig = config.XBoxLiveConfig;
 
-                var titleInfo = await _xApiService.GetMarketplaceAsync(xConfig.TitleId, xConfig.XApiKey, xConfig.Locale);
+                var titleInfo = await _xApiService.GetMarketplaceAsync(xConfig.TitleId, xConfig.XApiKey);
 
                 return new TitleInfo
                 {
+                    Platform = ActiveConfig.XBoxLive,
                     TitleId = xConfig.TitleId,
                     ProductTitle = titleInfo.Products.FirstOrDefault()?.LocalizedProperties?.FirstOrDefault()?.ProductTitle ?? "Unknown",
                     ProductDescription = titleInfo.Products.FirstOrDefault()?.LocalizedProperties?.FirstOrDefault()?.ProductDescription ?? "-",
@@ -52,14 +53,16 @@ namespace TwitchAchievementTrackerBackend.Controllers
             {
                 var steamConfig = config.SteamConfig;
 
+                var libraryUrlTask = _steamApiService.GetLibraryTileImage(long.Parse(steamConfig.AppId));
                 var titleInfo = await _steamApiService.GetStoreDetails(uint.Parse(steamConfig.AppId));
 
                 return new TitleInfo
                 {
+                    Platform = ActiveConfig.Steam,
                     TitleId = titleInfo.SteamAppid.ToString(),
                     ProductTitle = titleInfo.Name,
                     ProductDescription = titleInfo.ShortDescription,
-                    LogoUri = titleInfo.HeaderImage.ToString(),
+                    LogoUri =  (await libraryUrlTask) ?? titleInfo.HeaderImage.ToString(),
                 };
             }
 
@@ -75,7 +78,7 @@ namespace TwitchAchievementTrackerBackend.Controllers
         [HttpGet("/api/title/xapi/search/{query}")]
         public async Task<IEnumerable<TitleInfo>> SearchXApiTitleInfo(string query, string xApiKey = null)
         {
-            XApiConfiguration xConfig = null;
+            XApiConfiguration xConfig = new XApiConfiguration();
             if (this.HasExtensionConfiguration())
             {
                 var config = this.GetExtensionConfiguration();
@@ -119,6 +122,7 @@ namespace TwitchAchievementTrackerBackend.Controllers
                 return searchResult.Products.Select(product =>
                     new TitleInfo
                     {
+                        Platform = ActiveConfig.XBoxLive,
                         TitleId = product.AlternateIds?.FirstOrDefault(id => id.IdType == "XboxTitleId")?.Value ?? "",
                         ProductTitle = product.LocalizedProperties?.FirstOrDefault()?.ProductTitle ?? "Unknown",
                         ProductDescription = product.LocalizedProperties?.FirstOrDefault()?.ProductDescription ?? "-",
@@ -136,11 +140,6 @@ namespace TwitchAchievementTrackerBackend.Controllers
 
             if (config.ActiveConfig == ActiveConfig.XBoxLive)
             {
-                if (string.IsNullOrEmpty(config.XBoxLiveConfig.TitleId))
-                {
-
-                }
-
                 var achievements = await _xApiService.GetAchievementsAsync(config.XBoxLiveConfig);
                 var stateSummary = achievements.GroupBy(a => a.ProgressState).ToDictionary(a => a.Key, a => a.Count());
 
@@ -153,8 +152,8 @@ namespace TwitchAchievementTrackerBackend.Controllers
                     CurrentPoints = achievements
                         .Where(a => a.ProgressState == ProgressState.Achieved)
                         .SelectMany(a => a.Rewards.Where(r => r.Type == "Gamerscore"))
-                        .Sum(r => r.Value ?? 0),
-                    TotalPoints = achievements.SelectMany(a => a.Rewards.Where(r => r.Type == "Gamerscore")).Sum(r => r.Value ?? 0)
+                        .Sum(r => r.Value),
+                    TotalPoints = achievements.SelectMany(a => a.Rewards.Where(r => r.Type == "Gamerscore")).Sum(r => r.Value)
                 };
             }
             else if (config.ActiveConfig == ActiveConfig.Steam)
