@@ -3,15 +3,15 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using TwitchAchievementTrackerBackend.Model;
 using TwitchAchievementTrackerBackend.Services;
-using TwitchAchievementTracker_v1;
+using TwitchAchievementTracker.Flatbuffer.v1;
+using FlatSharp;
 
 namespace TwitchAchievementTrackerBackend.Extensions
 {
     public static class Compatibilityv001Extensions
     {
-        public static byte[] EncodeConfigurationToken_v1(this ConfigurationTokenService service, ExtensionConfiguration configuration)
+        public static byte[] EncodeConfigurationToken_v1(this ConfigurationTokenService service, Model.ExtensionConfiguration configuration)
         {
             byte[] payload = service.SerializeConfigurationToken_v1(configuration);
 
@@ -19,100 +19,103 @@ namespace TwitchAchievementTrackerBackend.Extensions
             return service.Encrypt(payload);
         }
 
-        public static byte[] SerializeConfigurationToken_v1(this ConfigurationTokenService service, ExtensionConfiguration configuration)
+        public static byte[] SerializeConfigurationToken_v1(this ConfigurationTokenService service, Model.ExtensionConfiguration configuration)
         {
             // Generate the Flatbuffer token payload
             var builder = new FlatBufferBuilder(512);
 
-            Offset<TwitchAchievementTracker_v1.Configuration> configurationOffset;
-            var version = builder.CreateString(configuration.Version);
+            var fbConfiguration = new TwitchAchievementTracker.Flatbuffer.v1.Configuration();
+            fbConfiguration.version = configuration.Version;
 
+            
             var xConfig = configuration.XBoxLiveConfig as Model.XApiConfiguration;
             var steamConfig = configuration.SteamConfig as Model.SteamConfiguration;
 
             switch (configuration.ActiveConfig)
             {
-                case ActiveConfig.XBoxLive:
-                    var apiKeyString = builder.CreateString(xConfig.XApiKey);
-                    var xuid64 = string.IsNullOrEmpty(xConfig.StreamerXuid) ? 0 : UInt64.Parse(xConfig.StreamerXuid);
-                    var titleId32 = string.IsNullOrEmpty(xConfig.TitleId) ? 0 : UInt32.Parse(xConfig.TitleId);
-                    var liveLocale = builder.CreateString(xConfig.Locale ?? "en-US");
+                case Model.ActiveConfig.XBoxLive:
 
-                    var xConfigOffset = TwitchAchievementTracker_v1.XApiConfiguration.CreateXApiConfiguration(builder, apiKeyString, xuid64, titleId32, liveLocale);
-                    configurationOffset = TwitchAchievementTracker_v1.Configuration.CreateConfiguration(builder, version, TwitchAchievementTracker_v1.PlatformConfiguration.XApiConfiguration, xConfigOffset.Value);
+                    var flatBufferXBoxConfig = new XApiConfiguration();
+                    flatBufferXBoxConfig.xApiKey = xConfig.XApiKey;
+                    flatBufferXBoxConfig.streamerXuid = string.IsNullOrEmpty(xConfig.StreamerXuid) ? 0 : UInt64.Parse(xConfig.StreamerXuid);
+                    flatBufferXBoxConfig.titleId = string.IsNullOrEmpty(xConfig.TitleId) ? 0 : UInt32.Parse(xConfig.TitleId);
+                    flatBufferXBoxConfig.locale = xConfig.Locale ?? "en-US";
+                    fbConfiguration.config = new PlatformConfiguration(flatBufferXBoxConfig);
 
                     break;
-                case ActiveConfig.Steam:
+                case Model.ActiveConfig.Steam:
                     var webKeyString = builder.CreateString(steamConfig.WebApiKey);
                     var steamId = UInt64.Parse(steamConfig.SteamId);
                     var appId = steamConfig.AppId;
                     var steamLocale = builder.CreateString(steamConfig.Locale ?? "english");
+                    var flatBufferSteamConfig = new SteamConfiguration();
+                    flatBufferSteamConfig.webApiKey = steamConfig.WebApiKey;
+                    flatBufferSteamConfig.steamId = string.IsNullOrEmpty(steamConfig.SteamId) ? 0 : UInt64.Parse(steamConfig.SteamId);
+                    flatBufferSteamConfig.appId = string.IsNullOrEmpty(steamConfig.AppId) ? 0 : UInt32.Parse(steamConfig.AppId);
+                    fbConfiguration.config = new PlatformConfiguration(flatBufferSteamConfig);
 
-                    var steamConfigOffset = TwitchAchievementTracker_v1.SteamConfiguration.CreateSteamConfiguration(builder, webKeyString, uint.Parse(appId), steamId, steamLocale);
-                    configurationOffset = TwitchAchievementTracker_v1.Configuration.CreateConfiguration(builder, version, TwitchAchievementTracker_v1.PlatformConfiguration.SteamConfiguration, steamConfigOffset.Value);
                     break;
                 default:
                     throw new NotSupportedException("Unknown configuration platform");
             }
 
-            builder.Finish(configurationOffset.Value);
-            return builder.SizedByteArray();
+            int maxBytesNeeded = FlatBufferSerializer.Default.GetMaxSize(fbConfiguration);
+            byte[] outBuffer = new byte[maxBytesNeeded];
+            int bytesWritten = FlatBufferSerializer.Default.Serialize(fbConfiguration, outBuffer);
+
+            return new ArraySegment<byte>(outBuffer, 0, bytesWritten).ToArray();
         }
 
         // Legacy token decryption from published version v0.1.1
-        public static ExtensionConfiguration DecodeConfigurationToken_v1(this ConfigurationTokenService service, byte[] payload)
+        public static Model.ExtensionConfiguration DecodeConfigurationToken_v1(this ConfigurationTokenService service, byte[] payload)
         {
             // Decrypt the token
             byte[] decrypted = service.Decrypt(payload);
             return service.DeserializeConfigurationToken_v1(decrypted);
         }
 
-        public static ExtensionConfiguration DeserializeConfigurationToken_v1(this ConfigurationTokenService service, byte[] decrypted)
+        public static Model.ExtensionConfiguration DeserializeConfigurationToken_v1(this ConfigurationTokenService service, byte[] decrypted)
         {
+            var fbConfiguration = FlatBufferSerializer.Default.Parse<TwitchAchievementTracker.Flatbuffer.v1.Configuration>(decrypted);
 
-            IPlatformConfiguration platformConfiguration;
-            ActiveConfig activeConfig;
 
-            // Initialize Flatbuffer
-            var fbConfig = TwitchAchievementTracker_v1.Configuration.GetRootAsConfiguration(new ByteBuffer(decrypted));
-            switch (fbConfig.ConfigType)
+            SteamConfiguration steamFbConfig = null;
+            XApiConfiguration xApiFbConfig = null;
+
+            Model.ActiveConfig activeConfig;
+            if (fbConfiguration.config?.Kind == PlatformConfiguration.ItemKind.SteamConfiguration)
             {
-                case TwitchAchievementTracker_v1.PlatformConfiguration.XApiConfiguration:
-
-                    var xConfig = fbConfig.Config<TwitchAchievementTracker_v1.XApiConfiguration>().GetValueOrDefault();
-
-                    activeConfig = ActiveConfig.XBoxLive;
-                    platformConfiguration = new Model.XApiConfiguration
-                    {
-                        XApiKey = xConfig.XApiKey,
-                        StreamerXuid = xConfig.StreamerXuid == 0 ? null : xConfig.StreamerXuid.ToString(),
-                        TitleId = xConfig.TitleId == 0 ? null : xConfig.TitleId.ToString(),
-                        Locale = xConfig.Locale,
-                    };
-                    break;
-                case TwitchAchievementTracker_v1.PlatformConfiguration.SteamConfiguration:
-
-                    var steamConfig = fbConfig.Config<TwitchAchievementTracker_v1.SteamConfiguration>().GetValueOrDefault();
-
-                    activeConfig = ActiveConfig.Steam;
-                    platformConfiguration = new Model.SteamConfiguration
-                    {
-                        WebApiKey = steamConfig.WebApiKey,
-                        SteamId = steamConfig.SteamId == 0 ? null : steamConfig.SteamId.ToString(),
-                        AppId = steamConfig.AppId == 0 ? null : steamConfig.AppId.ToString(),
-                        Locale = steamConfig.Locale,
-                    };
-                    break;
-                default:
-                    throw new NotSupportedException("Unknown type");
+                activeConfig = Model.ActiveConfig.Steam;
+                steamFbConfig = fbConfiguration.config?.SteamConfiguration;
+            }
+            if (fbConfiguration.config?.Kind == PlatformConfiguration.ItemKind.XApiConfiguration)
+            {
+                activeConfig = Model.ActiveConfig.XBoxLive;
+                xApiFbConfig= fbConfiguration.config?.XApiConfiguration;
+            }
+            else
+            {
+                throw new NotSupportedException("Unknown configuration platform");
             }
 
-            return new ExtensionConfiguration
+            return new Model.ExtensionConfiguration
             {
-                Version = fbConfig.Version,
+                Version = fbConfiguration.version,
                 ActiveConfig = activeConfig,
-                XBoxLiveConfig = platformConfiguration as Model.XApiConfiguration,
-                SteamConfig = platformConfiguration as Model.SteamConfiguration
+                SteamConfig = steamFbConfig == null ? null : new Model.SteamConfiguration
+                {
+                    WebApiKey = steamFbConfig.webApiKey,
+                    SteamId = steamFbConfig.steamId == 0 ? null : steamFbConfig.steamId.ToString(),
+                    AppId = steamFbConfig.appId == 0 ? null : steamFbConfig.appId.ToString(),
+                    Locale = steamFbConfig.locale,
+                },
+                XBoxLiveConfig = xApiFbConfig == null ? null : new Model.XApiConfiguration
+                {
+                    XApiKey = xApiFbConfig.xApiKey,
+                    StreamerXuid = xApiFbConfig.streamerXuid == 0 ? null : xApiFbConfig.streamerXuid.ToString(),
+                    TitleId = xApiFbConfig.titleId == 0 ? null : xApiFbConfig.titleId.ToString(),
+                    Locale = xApiFbConfig.locale,
+                }
             };
         }
     }
