@@ -1,12 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Text.Json.Serialization;
+using System;
 
 namespace TwitchAchievementTrackerBackend.Model.XApi
 {
     public class XApiAchievementProgression
     {
+        [JsonPropertyName("requirements")]
+        public XApiRequirement[] Requirements { get; set; }
+
+        [JsonPropertyName("timeUnlocked")]
         public DateTimeOffset TimeUnlocked { get; set; }
     }
 }
