@@ -1,6 +1,11 @@
-﻿using Microsoft.Extensions.Caching.Memory;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Localization;
+using Microsoft.Extensions.Caching.Memory;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using System;
+using System.Globalization;
+using System.Linq;
 using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
@@ -31,8 +36,9 @@ namespace TwitchAchievementTrackerBackend.Services
         private readonly HttpClient _httpClient;
         private readonly XApiOptions _options;
         private readonly IMemoryCache _cache;
+        private readonly ILogger _logger;
 
-        public XApiService(IHttpClientFactory httpClientFactory, IOptions<XApiOptions> options, IMemoryCache memoryCache)
+        public XApiService(IHttpClientFactory httpClientFactory, IOptions<XApiOptions> options, IMemoryCache memoryCache, ILogger<XApiService> logger)
         {
             _httpClient = httpClientFactory.CreateClient();
             if (string.IsNullOrEmpty(options.Value.XApiKey))
@@ -40,10 +46,11 @@ namespace TwitchAchievementTrackerBackend.Services
                 throw new ArgumentNullException("Could not find XApi key in configuration");
             }
             _options = options.Value;
-            _httpClient.DefaultRequestHeaders.Add("X-AUTH", _options.XApiKey);
-            _httpClient.DefaultRequestHeaders.Add("Accept-Language", "en-US,en;q=0.5");
-            _httpClient.BaseAddress = new Uri("https://xapi.us/v2/");
+            _httpClient.DefaultRequestHeaders.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", _options.XApiKey);
+            _httpClient.DefaultRequestHeaders.Add("Accept-Language", "en-US,en");
+            _httpClient.BaseAddress = new Uri("https://xapi.us/api/");
             _cache = memoryCache;
+            _logger = logger;
         }
 
         public string GetCacheKey(string call, XApiConfiguration config)
@@ -51,12 +58,59 @@ namespace TwitchAchievementTrackerBackend.Services
             return $"{call}:{config.TitleId}:{config.StreamerXuid}:{config.Locale}";
         }
 
-        public async Task<XUIDInfo> GetApiKeyXuid(string xApiKey)
+        private void SaveRateLimit(string xApiKey, HttpResponseMessage response)
         {
-            var message = new HttpRequestMessage(HttpMethod.Get, $"accountXuid");
-            message.Headers.Add("X-AUTH", xApiKey);
+            if (response.Headers.Contains("X-RateLimit-Limit"))
+            {
+                var rateLimit = int.Parse(response.Headers.GetValues("X-RateLimit-Limit").FirstOrDefault() ?? "60");
+                var rateLimitRemaining = int.Parse(response.Headers.GetValues("X-RateLimit-Remaining").FirstOrDefault() ?? "60");
+                var rateLimitResetTime = response.Headers.Contains("X-RateLimit-Reset") ? int.Parse(response.Headers.GetValues("X-RateLimit-Reset").FirstOrDefault() ?? "3600") : 3600;
+                var rateLimitObj = new RateLimits
+                {
+                    HourlyLimit = rateLimit,
+                    Remaining = rateLimitRemaining,
+                    ResetTime = DateTimeOffset.UtcNow.AddSeconds(rateLimitResetTime),
+                };
+
+                _cache.Set($"ratelimit:{xApiKey}", rateLimitObj, rateLimitObj.ResetTime);
+            }
+        }
+
+        public RateLimits GetRateLimit(string xApiKey)
+        {
+            if (_cache.TryGetValue<RateLimits>($"ratelimit:{xApiKey}", out var rateLimit))
+            {
+                return rateLimit;
+            }
+            return new RateLimits
+            {
+                HourlyLimit = 60,
+                Remaining = 60,
+                ResetTime = DateTimeOffset.UtcNow
+            };
+        }
+
+        private void CheckRateLimit(string xApiKey)
+        {
+            if (_cache.TryGetValue<RateLimits>($"ratelimit:{xApiKey}", out var rateLimit))
+            {
+                if (rateLimit.Remaining < rateLimit.HourlyLimit * 0.1)
+                {
+                    _logger.LogWarning($"Nearing xApi rate limit with {rateLimit.Remaining}/{rateLimit.HourlyLimit}, reset in {rateLimit.ResetTime - DateTimeOffset.UtcNow}");
+                }
+            }
+        }
+
+        public async Task<XApiKeyProfile> GetApiKeyProfile(string xApiKey)
+        {
+            var message = new HttpRequestMessage(HttpMethod.Get, "/api/profile");
+            message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
             var response = await _httpClient.SendAsync(message);
             var responseBody = await response.Content.ReadAsStringAsync();
+
+            SaveRateLimit(xApiKey, response);
+
+            // Note: upon an invalid token, new xapi.us returns an HTML page
             if (!response.IsSuccessStatusCode)
             {
                 var errorMessage = JsonSerializer.Deserialize<XApiError>(responseBody, new JsonSerializerOptions
@@ -70,8 +124,12 @@ namespace TwitchAchievementTrackerBackend.Services
                 }
                 throw new XApiException(errorCode, errorMessage.Error_Message);
             }
+            else if (response.Content.Headers.ContentType.MediaType != "application/json")
+            {
+                throw new XApiException("XApiUnauthorizedError", "Invalid xapi.us token");
+            }
 
-            return JsonSerializer.Deserialize<XUIDInfo>(responseBody, new JsonSerializerOptions
+            return JsonSerializer.Deserialize<XApiKeyProfile>(responseBody, new JsonSerializerOptions
             {
                 PropertyNameCaseInsensitive = true
             });
@@ -86,11 +144,17 @@ namespace TwitchAchievementTrackerBackend.Services
 
                 if (config != null)
                 {
-                    message.Headers.Add("X-AUTH", config.XApiKey);
-                    message.Headers.Add("Accept-Language", $"{config.Locale};q=1.0");
+                    message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", config.XApiKey);
+                    if (config.Locale != null)
+                    {
+                        message.Headers.Add("Accept-Language", $"{config.Locale}");
+                    }
                 }
 
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(config.XApiKey, response);
+
                 response.EnsureSuccessStatusCode();
 
                 using (var responseStream = await response.Content.ReadAsStreamAsync())
@@ -113,14 +177,75 @@ namespace TwitchAchievementTrackerBackend.Services
 
             if (!_cache.TryGetValue<string>(cacheKey, out var result))
             {
-                var message = new HttpRequestMessage(HttpMethod.Get, $"xuid/{gamerTag}");
-                message.Headers.Add("X-AUTH", xApiKey);
+                var profile = await ProfileForGamertag(gamerTag, xApiKey);
+                result = profile.Id;
+                _cache.Set(cacheKey, result);
+            }
+
+            return result;
+        }
+
+        public async Task<XApiProfileForGamertag> ProfileForGamertag(string gamerTag, string xApiKey)
+        {
+            var cacheKey = $"pfg:{gamerTag}";
+
+            if (!_cache.TryGetValue<XApiProfileForGamertag>(cacheKey, out var result))
+            {
+                var message = new HttpRequestMessage(HttpMethod.Get, $"{gamerTag}/profile-for-gamertag");
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
 
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(xApiKey, response);
+
                 response.EnsureSuccessStatusCode();
 
-                result = await response.Content.ReadAsStringAsync();
-                _cache.Set(cacheKey, result);
+                using (var responseStream = await response.Content.ReadAsStreamAsync())
+                {
+                    result = await JsonSerializer.DeserializeAsync<XApiProfileForGamertag>(responseStream, new JsonSerializerOptions
+                    {
+                        PropertyNameCaseInsensitive = true
+                    });
+                    // Cache for 1 hour (configurable)
+                    _cache.Set(cacheKey, result, _options.StaticDataCacheTime);
+                }
+            }
+
+            return result;
+        }
+
+        public async Task<bool> PurgeCache(XApiConfiguration config)
+        {
+            var cacheKey = GetCacheKey($"achievements", config);
+            var wasCached = _cache.TryGetValue<XApiAchievement[]>(cacheKey, out var cachedAchievements);
+            if (wasCached)
+            {
+                _cache.Remove(cacheKey);
+            }
+
+            var reloaded = await GetAchievementsAsync(config);
+
+            return reloaded.Count(a => a?.ProgressState == ProgressState.Achieved) != cachedAchievements.Count(a => a?.ProgressState == ProgressState.Achieved);
+        }
+
+        public async Task<XApiProfile> GetProfile(string xuid, string xApiKey)
+        {
+            var cacheKey = $"profile:{xuid}";
+            if (!_cache.TryGetValue<XApiProfile>(cacheKey, out var result))
+            {
+                var message = new HttpRequestMessage(HttpMethod.Get, $"{xuid}/profile");
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
+                var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(xApiKey, response);
+
+                response.EnsureSuccessStatusCode();
+                using (var responseStream = await response.Content.ReadAsStreamAsync())
+                {
+                    result = await JsonSerializer.DeserializeAsync<XApiProfile>(responseStream);
+                    // Cache for 1 hour (configurable)
+                    _cache.Set(cacheKey, result, _options.StaticDataCacheTime);
+                }
             }
 
             return result;
@@ -132,8 +257,10 @@ namespace TwitchAchievementTrackerBackend.Services
             if (!_cache.TryGetValue<XApiGamerCard>(cacheKey, out var result))
             {
                 var message = new HttpRequestMessage(HttpMethod.Get, $"{xuid}/gamercard");
-                message.Headers.Add("X-AUTH", xApiKey);
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(xApiKey, response);
 
                 response.EnsureSuccessStatusCode();
                 using (var responseStream = await response.Content.ReadAsStreamAsync())
@@ -161,17 +288,25 @@ namespace TwitchAchievementTrackerBackend.Services
             if (! _cache.TryGetValue<XApiAchievement[]>(cacheKey, out var result))
             {
                 var message = new HttpRequestMessage(HttpMethod.Get, $"{xuid}/achievements/{titleId}");
-                message.Headers.Add("X-AUTH", config.XApiKey);
-                message.Headers.Add("Accept-Language", $"{config.Locale};q=1.0");
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", config.XApiKey);
+                message.Headers.Add("Accept-Language", $"{config.Locale}");
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(config.XApiKey, response);
 
                 response.EnsureSuccessStatusCode();
                 using (var responseStream = await response.Content.ReadAsStreamAsync())
                 {
-                    result = await JsonSerializer.DeserializeAsync<XApiAchievement[]>(responseStream, new JsonSerializerOptions
+                    var paginatedResult = await JsonSerializer.DeserializeAsync<XApiAchievementsResult>(responseStream, new JsonSerializerOptions
                     {
                         PropertyNameCaseInsensitive = true
                     });
+                    if (paginatedResult.PagingInfo.ContinuationToken != null)
+                    {
+                        _logger.LogWarning("Achievements for titleId {titleId} returned paginated results !", titleId);
+                    }
+                    // TODO: when xapi.us documents how to pass the continuation token, handle pagination properly
+                    result = paginatedResult.Achievements;
                     // Cache for 1 minute (configurable)
                     _cache.Set(cacheKey, result, _options.ResultCacheTime);
                 }
@@ -186,12 +321,15 @@ namespace TwitchAchievementTrackerBackend.Services
             if (!_cache.TryGetValue<XApiMarketplaceTitleInfo>(cacheKey, out var result))
             {
                 var message = new HttpRequestMessage(HttpMethod.Get, $"marketplace/show/{titleId}");
-                message.Headers.Add("X-AUTH", xApiKey);
-                if (! string.IsNullOrEmpty(locale))
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
+                if (!string.IsNullOrEmpty(locale))
                 {
-                    message.Headers.Add("Accept-Language", $"{locale};q=1.0");
+                    message.Headers.Add("Accept-Language", $"{locale}");
                 }
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(xApiKey, response);
+
                 response.EnsureSuccessStatusCode();
                 using (var responseStream = await response.Content.ReadAsStreamAsync())
                 {
@@ -213,12 +351,15 @@ namespace TwitchAchievementTrackerBackend.Services
             if (!_cache.TryGetValue<XApiTitleHistoryResult>(cacheKey, out var result))
             {
                 var message = new HttpRequestMessage(HttpMethod.Get, $"{xuid}/title-history");
-                message.Headers.Add("X-AUTH", xApiKey);
+                message.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue("Bearer", xApiKey);
                 if (!string.IsNullOrEmpty(locale))
                 {
-                    message.Headers.Add("Accept-Language", $"{locale};q=1.0");
+                    message.Headers.Add("Accept-Language", $"{locale}");
                 }
                 var response = await _httpClient.SendAsync(message);
+
+                SaveRateLimit(xApiKey, response);
+
                 response.EnsureSuccessStatusCode();
                 using (var responseStream = await response.Content.ReadAsStreamAsync())
                 {
